开户完成

+++ JRGC/NewSection/ViewAndController/Controller/MicroBank(微商银行模块)/ UCFMicroBankOpenAccount(微商开户流程)/Api/UCFMicroBankIdentifysendCodeInfoAPI.m
@@ -49,7 +49,7 @@
  */
 - (id)requestArgument {
     return @{
-             @"fromSite":[NSString stringWithFormat:@"%d",_accoutType],
+             @"fromSite":[NSString stringWithFormat:@"%ld",(long)_accoutType],
              @"destPhoneNo": _destPhoneNo,
              @"isVms" : _isVms,
              @"type" : _type
